renames hook file
