+++ Bookcatalog.WebReact/src/data/bookHooks.ts
